requirements.txt
fastapi
mysql-connector-python
bcrypt
uvicorn
pydantic
